(core) Update the Info icon to better match our icon set
Summary:
The Info icon is changing from a curly "i" to a sans-serif "i" (from the Nucleo
icon set). It is currently only used for certain toast notifications.

For column headers (which is what it's intended for), it should be rendered
smaller, e.g. with width and height of 10px or 12px.

Test Plan: No behavior changes; looks tested manually.

Reviewers: jarek

Reviewed By: jarek

Differential Revision: https://phab.getgrist.com/D3771