skip building test harness in docker image
A small test harness bundle was recently added that is breaking
the docker image build. It could be added to the docker image,
but that would introduce a bunch of extraneous test file
dependencies. So this tweaks the build to simply skip the test
bundle if its primary source file is not found.