Linkingstate Refactor and LinkInfo
Big refactor to LinkingState
    pulls out descriptive/user-facing info into LinkingState

Adds Linkinfo to SelectBy/"Selector For" UI in right panel

Bugfix: Disables linking from Attachment columns

Bugfix/Behavior change: changed linking with empty RefLists to better
match behavior of refs.

    (for context: Linking by a blank Ref filters to show records with a
blank value for that Ref)
    Previously this didn't work with RefLists. Linking from a blank
refList would show no records (except in some cases involving summary
tables)
    Fixed this so that linking by a blank val consistently means "show
all records where the corresponding col is blank"

Tweaks to buildViewSectionDom, widgetTitle DOM
previously the title dom element (testid('viewsection-title')) was
flex-grow 1, so all the blank space in the widget-title-row was part of
the title element (but only the actual title was clickable). This swaps
it so that the title doesn't grow, and instead the space is taken up by
a small div which stretches to fill available space. Makes it easier to
add other dom around the title