(core) Fixing bug with switching ai tier limit
Summary:
Radiobutton on the billing page was wrongly connected
to the computed observable, which resulted in an error when one
was switching the tier using this radio button directly.

This was issue only in Firefox, as Chrome was muting error events
from cross origin scripts - so Grist ignored such errors. Now this
is changed and those errors are visible to Grist.

Test Plan: Updated

Reviewers: georgegevoian

Reviewed By: georgegevoian

Subscribers: georgegevoian

Differential Revision: https://phab.getgrist.com/D4119